+++ crates/scheduler/src/event.rs
@@ -26,6 +26,8 @@ use crate::Scheduler;
 
 #[cfg(feature = "board-api-button")]
 pub mod button;
+#[cfg(feature = "internal-board-api-fingerprint")]
+pub mod fingerprint;
 #[cfg(feature = "board-api-gpio")]
 pub mod gpio;
 pub mod platform;
@@ -48,6 +50,8 @@ pub struct InstId;
 pub enum Key<B: Board> {
     #[cfg(feature = "board-api-button")]
     Button(button::Key<B>),
+    #[cfg(feature = "internal-board-api-fingerprint")]
+    Fingerprint(fingerprint::Key),
     #[cfg(feature = "board-api-gpio")]
     Gpio(gpio::Key<B>),
     Platform(platform::Key),
@@ -79,6 +83,14 @@ impl<'a, B: Board> From<&'a Event<B>> for Key<B> {
             #[cfg(feature = "board-api-button")]
             Event::Button(event) => {
                 or_unreachable!("applet-api-button", [event], Key::Button(event.into()))
+            }
+            #[cfg(feature = "internal-board-api-fingerprint")]
+            Event::Fingerprint(event) => {
+                or_unreachable!(
+                    "internal-applet-api-fingerprint",
+                    [event],
+                    Key::Fingerprint(event.into())
+                )
             }
             #[cfg(feature = "board-api-gpio")]
             Event::Gpio(event) => {
@@ -117,6 +129,8 @@ impl<B: Board> Key<B> {
         match self {
             #[cfg(feature = "board-api-button")]
             Key::Button(x) => x.disable(),
+            #[cfg(feature = "internal-board-api-fingerprint")]
+            Key::Fingerprint(x) => x.disable::<B>(),
             #[cfg(feature = "board-api-gpio")]
             Key::Gpio(x) => x.disable(),
             Key::Platform(x) => x.disable(),
@@ -164,6 +178,14 @@ pub fn process<B: Board>(scheduler: &mut Scheduler<B>, event: Event<B>) {
         #[cfg(feature = "board-api-button")]
         Event::Button(event) => {
             or_unreachable!("applet-api-button", [event], button::process(event, &mut params))
+        }
+        #[cfg(feature = "internal-board-api-fingerprint")]
+        Event::Fingerprint(event) => {
+            or_unreachable!(
+                "internal-applet-api-fingerprint",
+                [event],
+                fingerprint::process(event, &mut params, applet)
+            )
         }
         #[cfg(feature = "board-api-gpio")]
         Event::Gpio(_) => or_unreachable!("applet-api-gpio", [], gpio::process()),